フィードバックページの基本的なレイアウトを別コンポーネントに切り出す
